Integrate Operation registry in OperationManager
