+++ Database/EventStore/Internal/OperationManager.hs
@@ -19,29 +19,23 @@ import ClassyPrelude
 
 --------------------------------------------------------------------------------
 import Database.EventStore.Internal.Communication
-import Database.EventStore.Internal.Generator
 import Database.EventStore.Internal.Logger
-import Database.EventStore.Internal.Manager.Operation.Model
+import Database.EventStore.Internal.Manager.Operation.Registry
 import Database.EventStore.Internal.Messaging
 import Database.EventStore.Internal.Callback
 import Database.EventStore.Internal.Types
 
 --------------------------------------------------------------------------------
-type OpModel      = Model (IO ())
-type OpTransition = Transition (IO ())
-
---------------------------------------------------------------------------------
 data Internal =
   Internal { _logger  :: Logger
            , _mainBus :: Bus
-           , _ref     :: IORef OpModel
+           , _reg     :: Registry
            }
 
 --------------------------------------------------------------------------------
 operationManager :: Logger -> Settings -> Bus -> IO ()
 operationManager logger setts mainBus = do
-  gen      <- newGenerator
-  internal <- Internal logger mainBus <$> newIORef (newModel setts gen)
+  internal <- Internal logger mainBus <$> newRegistry setts mainBus
 
   subscribe mainBus (onInit internal)
   subscribe mainBus (onNew internal)
@@ -51,66 +45,23 @@ operationManager logger setts mainBus = do
 
 --------------------------------------------------------------------------------
 onInit :: Internal -> SystemInit -> IO ()
-onInit Internal{..} _ =
-  publish _mainBus (Initialized OperationManager)
+onInit Internal{..} _ = publish _mainBus (Initialized OperationManager)
 
 --------------------------------------------------------------------------------
 onNew :: Internal -> SubmitOperation -> IO ()
-onNew i@Internal{..} (SubmitOperation p op) = do
-  model <- readIORef _ref
-  let callback outcome =
-        case outcome of
-          Left e -> do
-            logFormat _logger Error "Operation error: {}" (Only $ Shown e)
-            reject p e
-          Right a -> fulfill p a
-
-  nextModel <- interpret i (pushOperation callback op model)
-  atomicWriteIORef _ref nextModel
+onNew Internal{..} (SubmitOperation cb op) = register _reg op cb
 
 --------------------------------------------------------------------------------
 onRecv :: Internal -> PackageReceived -> IO ()
-onRecv i@Internal{..} (PackageReceived pkg) = do
-  model <- readIORef _ref
-  for_ (submitPackage pkg model) $ \outcome -> do
-    nextModel <- interpret i outcome
-    atomicWriteIORef _ref nextModel
-
---------------------------------------------------------------------------------
-doAbort :: Internal -> IO ()
-doAbort i@Internal{..} = do
-  model     <- readIORef _ref
-  nextModel <- interpretAbort i (abort model)
-  atomicWriteIORef _ref nextModel
+onRecv Internal{..} (PackageReceived pkg) = handlePackage _reg pkg
 
 --------------------------------------------------------------------------------
 onShutdown :: Internal -> SystemShutdown -> IO ()
-onShutdown i@Internal{..} _ = do
+onShutdown Internal{..} _ = do
   logMsg _logger Info "Shutting down..."
-  doAbort i
+  abortPendingRequests _reg
   publish _mainBus (ServiceTerminated OperationManager)
 
 --------------------------------------------------------------------------------
 onCheck :: Internal -> Check -> IO ()
 onCheck _ _ = return ()
-
---------------------------------------------------------------------------------
-interpret :: Internal -> OpTransition -> IO OpModel
-interpret Internal{..} = go
-  where
-    go (Produce action nxt)  = action >> go nxt
-    go (Transmit pkg nxt)    = publish _mainBus (TcpSend pkg) >> go nxt
-    go (Await m)             = return m
-    go (NotHandled info nxt) = do
-      let node = masterInfoNodeEndPoints info
-      publish _mainBus (ForceReconnect node)
-      go nxt
-
---------------------------------------------------------------------------------
-interpretAbort :: Internal -> OpTransition -> IO OpModel
-interpretAbort Internal{..} = go
-  where
-    go (Produce action nxt) = action >> go nxt
-    go (Transmit _ nxt)     = go nxt
-    go (Await m)            = return m
-    go (NotHandled _ nxt)   = go nxt
